+++ src/util/logger.ts
@@ -1,3 +1,4 @@
+import { format } from "node:util";
 import { type LogOutputChannel, window } from "vscode";
 
 export class Logger {
@@ -23,6 +24,14 @@ export class Logger {
 
 	public static info(info: string): void {
 		Logger.getInstance().out.info(info);
+	}
+
+	public static debug(debug: string, ...args: (string | number)[]): void {
+		Logger.getInstance().out.debug(format(debug, ...args));
+	}
+
+	public static trace(debug: string, ...args: (string | number)[]): void {
+		Logger.getInstance().out.trace(format(debug, ...args));
 	}
 
 	public dispose(): void {